+++ src/HealthCloudClient/Models/SleepSegment.cs
@@ -15,7 +15,7 @@ namespace HealthCloudClient
         /// <value>The sleep state = ['Unknown', 'UndifferentiatedSleep', 'RestlessSleep', 'RestfulSleep']</value>
         public string SleepType { get; set; }
         /// <value>The unique identifier of the segment</value>
-        public int? SegmentId { get; set; }
+        public long? SegmentId { get; set; }
         /// <value>The start time of the segment</value>
         public DateTime? StartTime { get; set; }
         /// <value>The end time of the segment</value>
